Search

+++ server/routers/order.js
@@ -17,7 +17,7 @@ module.exports = (passport, ormModels, sequelize) => {
 
   router.post("/create", async (req, res) => {
     const user = req.user;
-    console.log('+++++++++++++++');
+    console.log("+++++++++++++++");
     console.log(user);
     if (user && user.id) {
       const cart = await Cart.findOne({ where: { userId: user.id } });
@@ -34,9 +34,21 @@ module.exports = (passport, ormModels, sequelize) => {
             "product"."description" AS "product.description",
              "product"."image" AS "product.image",
               "product"."price" AS "orderedPrice",
-               "product"."stock" AS "product.stock" FROM "items" AS "items"
-                LEFT OUTER JOIN "products" AS "product" ON "items"."product_id" = "product"."id" WHERE "items"."cart_id" = ${cartId};
+               "product"."stock" AS "product.stock",
+                "product"."stock" AS "leftStock"
+                  FROM "items" AS "items"
+                  LEFT OUTER JOIN "products" AS "product" ON "items"."product_id" = "product"."id" WHERE "items"."cart_id" = ${cartId};
       `);
+      for (let i = 0; i < items[0].length; i++) {
+        const stock = items[0][i].stock;
+        const leftStock = items[0][i].leftStock;
+        if (stock > leftStock) {
+          return res.status(422).send({
+            error:
+              "Някой от продуктите във вашата количка са вече изчерпани, моля проверете количката си!"
+          });
+        }
+      }
       const order = await Order.findOne({
         where: { userId: user.id, status: "New" }
       });
@@ -55,25 +67,35 @@ module.exports = (passport, ormModels, sequelize) => {
         for (let i = 0; i < items[0].length; i++) {
           const productId = items[0][i].productId;
           const stock = items[0][i].stock;
+          const leftStock = items[0][i].leftStock;
           const orderedPrice = items[0][i].orderedPrice;
-          await OrderedItem.create({
+          const orderedItem = await OrderedItem.create({
             orderId: newOrderId,
             productId,
             stock,
             orderedPrice
           });
+          if (orderedItem) {
+            const product = await Product.update(
+              { stock: leftStock - stock },
+              { where: { id: productId } }
+            );
+          }
         }
       }
 
       //console.log(items);
       //console.log(cart);
       return res.sendStatus(200).json(items[0]);
-    } else res.sendStatus(403);
+    } else
+      res
+        .status(403)
+        .send({ error: "Имаше проблем със създаването на вашата поръчка." });
   });
 
   router.get("/get", async (req, res) => {
     const user = req.user;
-    console.log('--------------------');
+    console.log("--------------------");
     console.log(user);
     if (user && user.id) {
       const order = await Order.findOne({
@@ -126,10 +148,9 @@ module.exports = (passport, ormModels, sequelize) => {
           { where: { id: order.dataValues.id } }
         );
         if (updatedOrder) {
-          const cart = await Cart.findOne({where:{userId:user.id}});
-          if(cart)
-            await Cart.destroy({ where: { userId: user.id } });
-          console.log(cart)
+          const cart = await Cart.findOne({ where: { userId: user.id } });
+          if (cart) await Cart.destroy({ where: { userId: user.id } });
+          console.log(cart);
           if (cart)
             await Item.destroy({ where: { cartId: cart.dataValues.id } });
           res.sendStatus(200);
